+++ src/config.ts
@@ -12,16 +12,13 @@ export interface Config {
   actions: any[];
 }
 
-const repo = process.env.TRAVIS_REPO_SLUG;
-const pull = process.env.TRAVIS_PULL_REQUEST;
-
 export const getConfig = async (configLocation: string = '.lemmy.json'): Promise<Config> => {
   const file = await readFile(configLocation, 'utf-8');
   const config: Config = {
     git: {
-      baseBranch: 'master',
-      repo,
-      pull,
+      baseBranch: process.env.TRAVIS_PULL_REQUEST_BRANCH || 'master',
+      repo: process.env.TRAVIS_REPO_SLUG,
+      pull: process.env.TRAVIS_PULL_REQUEST,
     },
     message: {
       github: process.env.GITHUB_TOKEN,